requirements.txt
alembic==1.7.7
anyio==3.6.2
asgiref==3.4.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
cryptography==2.2.2
Deprecated==1.2.13
distlib==0.3.6
docopt==0.6.2
fastapi==0.83.0
filelock==3.4.1
greenlet==2.0.0
h11==0.13.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
jsonschema==3.2.0
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pipenv==2022.4.20
pipreqs==0.4.11
platformdirs==2.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
redis==4.3.4
requests==2.28.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.44
sqlalchemyseed==1.0.6
starlette==0.19.1
typing_extensions==4.1.1
urllib3==1.26.15
uvicorn==0.16.0
virtualenv==20.16.6
virtualenv-clone==0.5.7
wrapt==1.14.1
xlrd==1.2.0
yarg==0.1.9
zipp==3.6.0
